Extracts Business out to a separate class
